refactor: 컨트롤러 별로 테스트 분리
